python/requirements.txt
pybind11
numpy
sounddevice
mido
unittest
python-rtmidi